Refactor: Extract new mod `channel/open_channel.rs`
